+++ src/Symfony/Component/Serializer/CHANGELOG.md
@@ -22,6 +22,7 @@ CHANGELOG
  * Add `LoaderChainAwareInterface` so that loaders in a `LoaderChain` can defer work until all loaders have run; `LoaderChain` then validates discriminator maps declared in any format
  * Resolve generic template types during denormalization, e.g. `T` in a `Box<Circle>`-typed property is denormalized as `Circle`
  * Add an optional `$mappedClasses` argument to `Mapping\Loader\LoaderChain::__construct()`, to declare which classes each loader maps
+ * Trigger a deprecation when denormalizing a property from its PHP name while a name converter maps it to another key
 
 8.1
 ---